+++ jOOU/src/test/java/org/joou/test/UByteTest.java
@@ -35,16 +35,16 @@
  */
 package org.joou.test;
 
+import static junit.framework.Assert.assertEquals;
+import static junit.framework.Assert.assertTrue;
+import static junit.framework.Assert.fail;
+import static org.joou.Unsigned.ubyte;
+
 import java.io.ByteArrayInputStream;
 import java.io.ByteArrayOutputStream;
 import java.io.IOException;
 import java.io.ObjectInputStream;
 import java.io.ObjectOutputStream;
-
-import static junit.framework.Assert.assertEquals;
-import static junit.framework.Assert.assertTrue;
-import static junit.framework.Assert.fail;
-import static org.joou.Unsigned.ubyte;
 
 import org.joou.UByte;
 import org.junit.Test;
@@ -52,110 +52,82 @@ import org.junit.Test;
 public class UByteTest {
 
     @Test
-    public void testValueOfByte()
-    {
-        for(int i=Byte.MIN_VALUE;i<=Byte.MAX_VALUE;i++)
-        assertEquals(i&0xFF, ubyte((byte)i).intValue());
+    public void testValueOfByte() {
+        for (int i = Byte.MIN_VALUE; i <= Byte.MAX_VALUE; i++)
+            assertEquals(i & 0xFF, ubyte((byte) i).intValue());
     }
 
     @Test
-    public void testValueOfByteCaching()
-    {
-        for(int i=Byte.MIN_VALUE;i<=Byte.MAX_VALUE;i++)
-        {
-            UByte a = ubyte((byte)i);
-            UByte b = ubyte((byte)i);
+    public void testValueOfByteCaching() {
+        for (int i = Byte.MIN_VALUE; i <= Byte.MAX_VALUE; i++) {
+            UByte a = ubyte((byte) i);
+            UByte b = ubyte((byte) i);
             assertTrue(a == b);
         }
     }
 
     @Test
-    public void testValueOfShort()
-    {
-        for(int i=UByte.MIN_VALUE;i<=UByte.MAX_VALUE;i++)
-            assertEquals(i&0xFF, ubyte((short)i).intValue());
+    public void testValueOfShort() {
+        for (int i = UByte.MIN_VALUE; i <= UByte.MAX_VALUE; i++)
+            assertEquals(i & 0xFF, ubyte((short) i).intValue());
     }
 
     @Test
-    public void testValueOfShortInvalid()
-    {
-        try
-        {
-            ubyte((short)((int)(UByte.MIN_VALUE)-1));
+    public void testValueOfShortInvalid() {
+        try {
+            ubyte((short) ((UByte.MIN_VALUE) - 1));
             fail();
         }
-        catch(NumberFormatException e)
-        {
-        }
-        try
-        {
-            ubyte((short)((int)(UByte.MAX_VALUE)+1));
+        catch (NumberFormatException e) {}
+        try {
+            ubyte((short) ((UByte.MAX_VALUE) + 1));
             fail();
         }
-        catch(NumberFormatException e)
-        {
-        }
+        catch (NumberFormatException e) {}
     }
 
     @Test
-    public void testValueOfInt()
-    {
-        for(int i=UByte.MIN_VALUE;i<=UByte.MAX_VALUE;i++)
-            assertEquals(i&0xFF, ubyte((int)i).intValue());
+    public void testValueOfInt() {
+        for (int i = UByte.MIN_VALUE; i <= UByte.MAX_VALUE; i++)
+            assertEquals(i & 0xFF, ubyte(i).intValue());
     }
 
     @Test
-    public void testValueOfIntInvalid()
-    {
-        try
-        {
-            ubyte((int)(UByte.MIN_VALUE)-1);
+    public void testValueOfIntInvalid() {
+        try {
+            ubyte((UByte.MIN_VALUE) - 1);
             fail();
         }
-        catch(NumberFormatException e)
-        {
-        }
-        try
-        {
-            ubyte((int)(UByte.MAX_VALUE)+1);
+        catch (NumberFormatException e) {}
+        try {
+            ubyte((UByte.MAX_VALUE) + 1);
             fail();
         }
-        catch(NumberFormatException e)
-        {
-        }
+        catch (NumberFormatException e) {}
     }
 
     @Test
-    public void testValueOfLong()
-    {
-        for(int i=UByte.MIN_VALUE;i<=UByte.MAX_VALUE;i++)
-            assertEquals(i&0xFF, ubyte((long)i).intValue());
+    public void testValueOfLong() {
+        for (int i = UByte.MIN_VALUE; i <= UByte.MAX_VALUE; i++)
+            assertEquals(i & 0xFF, ubyte((long) i).intValue());
     }
 
     @Test
-    public void testValueOfLongInvalid()
-    {
-        try
-        {
-            ubyte((long)(UByte.MIN_VALUE)-1);
+    public void testValueOfLongInvalid() {
+        try {
+            ubyte((long) (UByte.MIN_VALUE) - 1);
             fail();
         }
-        catch(NumberFormatException e)
-        {
-        }
-        try
-        {
-            ubyte((long)(UByte.MAX_VALUE)+1);
+        catch (NumberFormatException e) {}
+        try {
+            ubyte((long) (UByte.MAX_VALUE) + 1);
             fail();
         }
-        catch(NumberFormatException e)
-        {
-        }
+        catch (NumberFormatException e) {}
     }
 
     @Test
-    public void testSerializeDeserialize() throws ClassNotFoundException, IOException
-    {
+    public void testSerializeDeserialize() throws ClassNotFoundException, IOException {
         ByteArrayOutputStream baos = new ByteArrayOutputStream();
         ObjectOutputStream oos = new ObjectOutputStream(baos);
         ByteArrayInputStream bais;
@@ -170,11 +142,11 @@ public class UByteTest {
         bais = new ByteArrayInputStream(baos.toByteArray());
         ois = new ObjectInputStream(bais);
         o = ois.readObject();
-        if(!(o instanceof UByte))
-                fail();
-        actual = (UByte)o;
-        assertEquals(expected, actual);	// same value
-        assertTrue(expected == actual);	// identical objects
+        if (!(o instanceof UByte))
+            fail();
+        actual = (UByte) o;
+        assertEquals(expected, actual); // same value
+        assertTrue(expected == actual); // identical objects
     }
 
 }
